requirements.txt
kafka-python==2.0.2
faker==19.13.0
tqdm==4.66.1 